\name{A_old}
\docType{data}
\alias{A_old}
\title{TF-gene regulation strength matrix sampled from the previous round}
\description{
  A matrix of TF-gene regulation strength with genes as rows and TFs as columns, sampled from the previous round. During the Gibbs sampling process, this matrix is used as prior for a new round of regulation strength sampling.
}
\usage{A_old}
\format{numeric matrix}
\keyword{Regulation_Strength}
