\name{base_line_old}
\docType{data}
\alias{base_line_old}
\title{Gene baseline expression sampled from the previous round.}
\description{
  A vector of baseline expression for all genes, sampled from the previous round. During the Gibbs Samplig process, this is used as a prior for a new round of gene baseline expression sampling. 
}
\usage{base_line_old}
\format{numeric vector}
\keyword{Baseline}
