% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_sampling.R
\name{baseline_sampling}
\alias{baseline_sampling}
\title{Gene Baseline Expression Sampling Function}
\usage{
baseline_sampling(Y, C, A, X, base_line_old, C_prior, sigma_noise, sigma_A,
  sigma_baseline, sigma_X)
}
\arguments{
\item{Y}{gene expression data matrix}

\item{C}{sampled TF-gene binding network}

\item{A}{sampled regulatory strength matrix}

\item{X}{sampled transcription factor activity matrix}

\item{base_line_old}{prior gene expression baseline activity}

\item{C_prior}{prior TF-gene binding network}

\item{sigma_noise}{variance of gene expression fitting residuals}

\item{sigma_A}{variance of regulatory strength}

\item{sigma_baseline}{variance of gene expression baseline activity}

\item{sigma_X}{variance of transcription factor activity}
}
\description{
Function 'baseline_sampling' estimates a baseline expression for each gene,
according to a posterior Gaussian distribution.
}
\keyword{Baseline}
\keyword{sampling}
