\name{C_old}
\docType{data}
\alias{C_old}
\title{TF-gene binding network sampled from the previous round}
\description{
  A matrix of TF-gene binding network sampled from the previous round, with each unit either 1 (binding) or 0 (non-binding). During the Gibbs sampling process, this is used as a prior for a new round of binding network sampling.
}
\usage{C_old}
\format{numeric matrix}
\keyword{Binding}
