\name{C_prior}
\docType{data}
\alias{C_prior}
\title{Prior TF-gene binding network}
\description{
  A matrix of prior TF-gene binding events, with each unit either 1 (binding) or 0 (non-binding). Such a prior network can be obtained from TF-gene binding database, motif searching, ChIP-seq peaks or ATAC-seq peaks.
}
\usage{C_prior}
\format{numeric matrix}
\keyword{Binding}
