% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C_sampling_cluster.R
\name{C_sampling_cluster}
\alias{C_sampling_cluster}
\title{cis-Regulatory Module Sampling Function}
\usage{
C_sampling_cluster(Y, C_old, A_old, X_old, base_line_old, C_prior, sigma_noise,
  sigma_A, sigma_baseline, sigma_X, BICORN_input)
}
\arguments{
\item{Y}{gene expression data matrix}

\item{C_old}{TF-gene binding network sampled from the previous round}

\item{A_old}{regulatory strength matrix sampled from the previous round}

\item{X_old}{transcription factor activity matrix sampled from the previous round}

\item{base_line_old}{gene expression baseline activity sampled from the previous round}

\item{C_prior}{prior TF-gene binding network}

\item{sigma_noise}{variance of gene expression fitting residuals}

\item{sigma_A}{variance of regulatory strength}

\item{sigma_baseline}{variance of gene expression baseline activity}

\item{sigma_X}{variance of transcription factor activity}

\item{BICORN_input}{this list structure contains TF symbols, Gene symbols and candidate modules}
}
\description{
Function 'C_sampling_cluster' samples a candidate cis-regulatory module for each gene,
according to a discrete posterior probability distribution.
}
\keyword{Module}
\keyword{sampling}
