\name{Exp_data}
\docType{data}
\alias{Exp_data}
\title{Gene expression data}
\description{
  A matrix of normalized gene expression data with genes as rows and samples as columns. The gene expression data can be either time-course data measured under multiple time points or steady state data generated from at least two different conditions.
}
\usage{Exp_data}
\format{numeric matrix}
\keyword{Gene_expression}
