% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmanoise_sampling.R
\name{sigmanoise_sampling}
\alias{sigmanoise_sampling}
\title{Fitting Residule Variance Sampling Function}
\usage{
sigmanoise_sampling(Y, C, A, X, base_line, C_prior, sigma_A, sigma_baseline,
  sigma_X, alpha, beta)
}
\arguments{
\item{Y}{gene expression data matrix}

\item{C}{sampled TF-gene binding network}

\item{A}{sampled regulatory strength matrix}

\item{X}{sampled transcription factor activity matrix}

\item{base_line}{sampled gene expression baseline activity}

\item{C_prior}{prior TF-gene binding network}

\item{sigma_A}{variance of regulatory strength}

\item{sigma_baseline}{variance of gene expression baseline activity}

\item{sigma_X}{variance of transcription factor activity}

\item{alpha}{hyper-parameter for inverse-gamma distribution}

\item{beta}{hyper-parameter for inverse-gamma distribution}
}
\description{
Function 'sigmanoise_sampling' estimates the variance of overal gene expression fitting residuals,
according to an inverse-gamma distribution.
}
\keyword{Noise}
\keyword{sampling}
