\name{X_old}
\docType{data}
\alias{X_old}
\title{Transcription factr activity matrix sampled from the previous round}
\description{
  A matrix of hidden transcription factr activity estimated from gene expression data, with transcription factrs as rows and samples as columns, sampled from the previous round. During the Gibbs sampling process, this is used as a prior for a new round of transcription factor activity sampling.
}
\usage{X_old}
\format{numeric matrix}
\keyword{TFA}
