% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X_sampling.R
\name{X_sampling}
\alias{X_sampling}
\title{Transcription Factor Activity Sampling Function}
\usage{
X_sampling(Y, C, A, X_old, base_line, C_prior, sigma_noise, sigma_A,
  sigma_baseline, sigma_X)
}
\arguments{
\item{Y}{gene expression data matrix}

\item{C}{sampled TF-gene binding network}

\item{A}{sampled regulatory strength matrix}

\item{X_old}{sampled transcription factor activity matrix from the previous round}

\item{base_line}{sampled gene expression baseline activity}

\item{C_prior}{prior TF-gene binding network}

\item{sigma_noise}{variance of gene expression fitting residuals}

\item{sigma_A}{variance of regulatory strength}

\item{sigma_baseline}{variance of gene expression baseline activity}

\item{sigma_X}{variance of transcription factor activity}
}
\description{
Function 'X_sampling' estimates the hidden activities of each transcription factor,
according to a posterior Gaussian random process.
}
\keyword{TFA}
\keyword{sampling}
