\name{Y}
\docType{data}
\alias{Y}
\title{Gene expression data used for module inference}
\description{
  A matrix of normalized gene expression for common genes of prior binding input and gene expression input, with genes as rows and samples as columns. Y is the matrix used for cis-regulatory mudole inference.
}
\usage{Y}
\format{numeric matrix}
\keyword{Gene_expression}
